Exile 1.0.4 compatibility and readme updates
2 construction overrides with changes.
This was to get it in the hands of all that use before the weekend.
Testing is still being done to ensure full function but should be fine
as these changes should not effect the vector!

btw, I noticed some added escape mission stuff in the keyUp file.
These changes were added to keep all building functions compatible with
your mission files and modes